Feat : Login 페이지 뷰 연동 완료
- login 페이지의 회원가입, 아이디 찾기 뷰 연동 완료
- 다른 템플릿을 사용해 html 파일 리펙터링 필요

+++ project/src/main/java/com/example/project/dto/FollowerDTO.java
@@ -11,7 +11,7 @@ import lombok.Setter;
 @NoArgsConstructor
 public class FollowerDTO {
 
-    private int seq;
+    private Long seq;
     private String follower_user_seq;
     private String following_user_seq;
     private String created_at;
